test: Remove readiness check from microscope
The readiness check is running a microscope command that takes several seconds
to complete. The interval is at 5 seconds. This delays the microscope pod from
being marked ready by ~10s.

It is unclear what the readiness command fullfils as it has no relation to the
actual microscope instance inside of the pod that will be run afterwards.

Signed-off-by: Thomas Graf <thomas@cilium.io>