test: Forcefully shut down microscope without any delays
Microscope does not seem to handle the SIGTERM signal properly. A likely cause
is the failure of Alpine linux to propagate signals to child processes which
prevents the sleep from being terminated. This causes kubectl to wait for the
gracePeriod to be complete before killing the container for real.

Just forcefully shut down the microscope pod as all of the data is already
extracted.

Signed-off-by: Thomas Graf <thomas@cilium.io>