daemon: increase function queue size to 1024 for all function queues
This allows for a lot of resources to be fetched before they are processed
from kube-apiserver. This allows for regeneration of endpoints that are
being restored from the host to not be blocked if the function queue is
filled up quickly, as the size is increased from 20 to 1024 for all
Kubernetes resources which Cilium watches for.

Signed-off by: Ian Vernon <ian@cilium.io>